backend/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

appengine-python-standard>=1.0.0
https://download.pytorch.org/whl/cpu/torch-1.10.1%2Bcpu-cp38-cp38-linux_x86_64.whl
google-cloud-firestore==2.5.3
Flask == 2.1.2
Flask_Cors == 3.0.10
Flask_Session == 0.4.0
gym == 0.21.0
numpy == 1.22.4
pandas == 1.2.4
requests == 2.27.1
six == 1.15.0
scikit-learn == 0.24.1
stable_baselines3 == 1.5.0